% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_controls}
\alias{cps_controls}
\title{CPS controls (15,992 observations)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15992 rows and 11 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
cps_controls
}
\description{
Non-experimental comparison data file constructed by Lalonde from
the Current Population Survey.
}
\keyword{datasets}
