% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_controls2}
\alias{cps_controls2}
\title{CPS2 controls (2,369 observations)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2369 rows and 11 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
cps_controls2
}
\description{
Further subsets created by Lalonde from the CPS non-experimental
comparison data file (\code{lalonde::cps_controls}).
}
\keyword{datasets}
