% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalonde-package.r
\docType{package}
\name{lalonde}
\alias{lalonde}
\alias{lalonde-package}
\title{lalonde.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\description{
An R data package for the Lalonde dataset.
}
