% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw}
\alias{nsw}
\title{NSW Data Files (Lalonde Sample)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 722 rows and 10 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
nsw
}
\description{
These files contain the treated and control units from the male
sub-sample from the National Supported Work Demonstration as used by Lalonde
in his paper.
}
\keyword{datasets}
