% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_dw}
\alias{nsw_dw}
\title{NSW Data Files (Dehejia-Wahha Sample)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 445 rows and 11 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
nsw_dw
}
\description{
Based on pre-intervention variables, Dehejia-Wahha extract a
further subset of Lalonde's NSW experimental data, a subset containing
information on RE74 (earnings in 1974).
}
\keyword{datasets}
