% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psid_controls}
\alias{psid_controls}
\title{PSID controls (2490 observations)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2490 rows and 11 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
psid_controls
}
\description{
Non-experimental comparison data file constructed by Lalonde
from the Population Survey of Income Dynamics.
}
\keyword{datasets}
