% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psid_controls2}
\alias{psid_controls2}
\title{PSID2 controls (253 observations)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 253 rows and 11 columns.}
\source{
\href{http://users.nber.org/~rdehejia/nswdata2.html}{Dehejia's NBER website.}
}
\usage{
psid_controls2
}
\description{
Further subsets created by Lalonde from the PSID non-experimental
comparison data file (\code{lalonde::psid_controls}).
}
\keyword{datasets}
